LayoutTests/storage/indexeddb/modern/deletedatabase-request-event-expected.txt

This test calls deleteDatabase on window.indexedDB and verifies that an event is fired at the request.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


ALERT: [object IDBOpenDBRequest]
ALERT: success
PASS successfullyParsed is true

TEST COMPLETE
